+++ mono_repo/lib/src/runner.dart
@@ -9,6 +9,7 @@ import 'commands/init.dart';
 import 'commands/presubmit.dart';
 import 'commands/pub.dart';
 import 'commands/travis.dart';
+import 'utils.dart';
 
 class MonoRepoRunner extends CommandRunner<Null> {
   MonoRepoRunner()
@@ -21,5 +22,9 @@ class MonoRepoRunner extends CommandRunner<Null> {
       new PubCommand(),
       new TravisCommand()
     ].forEach(addCommand);
+    argParser.addFlag(recursiveFlag,
+        help:
+            'Whether to recursively walk sub-directorys looking for packages.',
+        defaultsTo: false);
   }
 }